Test: Authentication api endpoints tested
