+++ package.json
@@ -5,7 +5,8 @@
   "main": "app.js",
   "scripts": {
     "start": "nodemon --experimental-modules --es-module-specifier-resolution=node app.js",
-    "test": "mocha"
+    "test:doc": "mocha 'test/**/*.js' --recursive --reporter mocha-junit-reporter --timeout 60000 --exit",
+    "test": "mocha 'test/**/*.js' --timeout 60000 --exit"
   },
   "author": "Lawrence Katuva",
   "license": "MIT",
@@ -18,7 +19,7 @@
     "express": "^4.17.1",
     "http-errors": "^1.8.0",
     "jsonwebtoken": "^8.5.1",
-    "mocha": "^9.1.1",
+    "mocha-junit-reporter": "^2.0.0",
     "mysql2": "^2.3.0",
     "nodemailer": "^6.6.3",
     "nodemon": "^2.0.12",
@@ -26,6 +27,7 @@
     "supertest": "^6.1.6"
   },
   "devDependencies": {
+    "mocha": "^9.1.1",
     "sequelize-cli": "^6.2.0"
   }
 }
